+++ src/DrugResponseModel.jl
@@ -25,13 +25,14 @@ include("allDrugs.jl")
 include("combination.jl")
 include("replicates.jl")
 include("sensitivity.jl")
+include("fitCombo.jl")
 include("figures/figureS2.jl")
 include("figures/figure3.jl")
 include("figures/figureS3.jl")
 include("figures/figure4.jl")
 include("figures/figureS4.jl")
 include("figures/figure5.jl")
-include("figures/figureS5.jl")
+# include("figures/figureS5.jl")
 
 export setup_data,
     load,
